+++ pytket/docs/changelog.rst
@@ -6,7 +6,7 @@ Changelog
 
 Minor new features:
 
-* Add new ``MaxNClRegPredicate`` that checks that there are at max n classical
+* Add new ``MaxNClRegPredicate`` that checks that there are at most n classical
   registers in the circuit.
 
 1.14.0 (April 2023)